chore(tx-backend): #410 typo in package name and use lombok for class constructor
