+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/submodel/infrastructure/repository/JpaSubmodelPayloadRepository.java
@@ -17,24 +17,11 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-package org.eclipse.tractusx.traceability.submodel.infrastructure.reposotory;
+package org.eclipse.tractusx.traceability.submodel.infrastructure.repository;
 
-import feign.Param;
-import feign.RequestLine;
-import org.eclipse.tractusx.traceability.common.config.SubmodelApiConfig;
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.RequestBody;
+import org.eclipse.tractusx.traceability.submodel.infrastructure.model.SubmodelPayloadEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-@FeignClient(
-        contextId = "submodelServerContextId",
-        name = "submodelServer",
-        url = "${feign.submodelApi.url}",
-        configuration = {SubmodelApiConfig.class}
-)
-public interface SubmodelServerApiClient {
-    @RequestLine("POST /api/submodel/data/{submodelId}")
-    void createSubmodel(@Param("submodelId") String submodelId, @RequestBody String payload);
+public interface JpaSubmodelPayloadRepository extends JpaRepository<SubmodelPayloadEntity, String> {
 
-    @RequestLine("GET /api/submodel/data/{submodelId}")
-    String getSubmodel(@Param("submodelId") String submodelId);
 }
