+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/integration/submodel/SubmodelServerClientTest.java
@@ -1,17 +1,15 @@
 package org.eclipse.tractusx.traceability.integration.submodel;
 
-import org.eclipse.tractusx.traceability.assets.domain.importpoc.repository.SubmodelPayloadRepository;
 import org.eclipse.tractusx.traceability.integration.IntegrationTestSpecification;
 import org.eclipse.tractusx.traceability.submodel.infrastructure.model.SubmodelEntity;
-import org.eclipse.tractusx.traceability.submodel.infrastructure.reposotory.JpaSubmodelRepository;
-import org.eclipse.tractusx.traceability.submodel.infrastructure.reposotory.SubmodelServerClient;
+import org.eclipse.tractusx.traceability.submodel.infrastructure.repository.JpaSubmodelRepository;
+import org.eclipse.tractusx.traceability.submodel.infrastructure.repository.SubmodelServerClient;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class SubmodelServerClientTest extends IntegrationTestSpecification {
